requirements.txt
# Cyclone v2 Project Requirements

    # --- Core & Scheduling ---
    python-dotenv>=1.0.0
    schedule>=1.2.0
    requests>=2.31.0 # Used for CryptoNews API and potentially others
    pandas>=2.2.0 # Ensure pandas version supports needed features
    numpy>=1.24.0
    pytz>=2023.3 # For timezone handling

    # --- Database ---
    sqlalchemy>=2.0.0 # Use modern SQLAlchemy
    # psycopg2-binary # Commented out - Not needed for Azure SQL
    pyodbc # Uncommented - Needed for Azure SQL/MS SQL Server

    # --- APIs ---
    python-binance>=1.0.17 # For Binance API
    pandas-ta # For Technical Indicators
    # cryptonews-api # Removed - No specific library found, using 'requests'
    praw>=7.7.0 # For Reddit API
    tweepy>=4.14.0 # For Twitter API v2
    python-decouple>=3.8
    aiohttp>=3.9.1
    asyncio>=3.4.3

    # --- ML & Sentiment ---
    transformers>=4.30.0 # Or transformers[tensorflow]
    torch # Or tensorflow
    scikit-learn>=1.3.0
    lightgbm>=4.0.0 # Or xgboost
    joblib # For saving/loading models
    huggingface-hub>=0.19.0

    # --- Dashboard ---
    dash>=3.0.0
    plotly>=6.0.0
    dash-bootstrap-components>=2.0.0 # For better dashboard layout/styling

    # --- Utilities ---
    # Add any other utility libraries needed
    tenacity>=8.2.0
    pytest>=7.0.0
    pytest-cov>=4.1.0
